+++ gitbook/other/change-log.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [8.6.0]
 
+### Fixed
+- Bumped Axios version to `0.30` to fix vulnerability `CVE-2025-27152`
+
 ## [8.5.0]
 
 ### Added