+++ gitbook/other/change-log.md
@@ -9,6 +9,38 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.10.0]
+
+### Fixed
+- The following vulnerabilities were resolved, but not applicable:
+  - Bumped Axios version to resolve `CVE-2026-40175` and `CVE-2025-62718`
+  - Bumped brace-expansion version to resolve `CVE-2026-33750`
+  - Bumped picomatch version to resolve `CVE-2026-33672` and `CVE-2026-33671`
+  - Bumped flatted version to resolve `CVE-2026-33228` and `CVE-2026-32141`
+
+### Added
+- `exh.users.getEmailTemplates` and `exh.users.setEmailTemplates` now support `_name` variants
+  - `activationEmailTemplateName`
+  - `reactivationEmailTemplateName`
+  - `passwordResetEmailTemplateName`
+  - `oidcUnlinkEmailTemplateName`
+  - `oidcUnlinkPinEmailTemplateName`
+  - `activationPinEmailTemplateName`
+  - `reactivationPinEmailTemplateName`
+  - `passwordResetPinEmailTemplateName`
+
+### Deprecated
+- `exh.users.getEmailTemplates` and `exh.users.setEmailTemplates` deprecates `_id` in favour of `_name` variants
+  - `activationEmailTemplateId`
+  - `reactivationEmailTemplateId` 
+  - `passwordResetEmailTemplateId`
+  - `oidcUnlinkEmailTemplateId`
+  - `oidcUnlinkPinEmailTemplateId`
+  - `activationPinEmailTemplateId`
+  - `reactivationPinEmailTemplateId`
+  - `passwordResetPinEmailTemplateId`
+
+
 ## [8.9.1]
 
 ### Fixed